virtxml: Errors loading the Osinfo DB are not fatal
An error will be reported even if loading partially worked (for
example when the system DB could be loaded, but the user DB failed
to load). Since the code will still work correctly even if the DB
could not be loaded at all, we can ignore the error when DB loading
failed. (code still works but virtxml cannot do a lot of useful stuff
as it needs an OS and an hypervisor).