Fix libvirt caps -> libosinfo platform short id mapping
The code was building an id starting with kvm- while libosinfo qemu
description uses qemu-kvm-. Also, starting from qemu 1.2.0, there is
no separate qemu-kvm tarball.
guess_platform_from_connect is starting to be a bit magic, it may
be better to add a <machine> attribute to libosinfo <platform>
description and to use this to improve the matching between
libosinfo data and libvirt caps.