Improve VM skeleton created by gvir_designer_domain_setup_guest()
Add various devices/configuration to libvirt XML config when
creating the VM. This configuration is generic enough that it
should be useful on all created VMs, that's why no public API is
added to set them up. However, they are split in several helpers
that can easily be exported if needed.

What this commit adds is:
- clock
- input device (a tablet which will act as a mouse)
- a console (Boxes is adding one, not sure if it's required...)

For maximum flexibility, we may want to let applications decide whether
they want a console or not.

In addition to these devices, gvir_designer_domain_setup_guest()
now disables power management, and sets the domain to stop on
power off/crash, and to restart on reboots.